Replace an AssetCard's (backend) asset, bugfixes
Summary:
To complete the templating update, we need to be able to replace a template asset with actual attached assets.
This is now done using a dropdown underneath the asset's "Content" settings. Disabled if no other assets fit the 'type'.

Also:

- Set DeleteAsset Folder to always open
- Back to MV -> "Back to Project"
- Filter non-ready assets from Scene Background dropdown
- Audio and Video components now actually reactive
- Removed unused import

Test Plan: Use a template, add an asset of corresponding test-type to resulting Project, replace asset in the placeholder asset's options

Reviewers: girish.vadlamudi, stephan.gensch

Reviewed By: girish.vadlamudi

Differential Revision: https://hive.vrgmnts.net/D1812