View mode playable if Story set to non-discoverable
Summary:
A super hacky way of checking whether the user is logged in and can access the Story in question: I'm calling API via one of the private calls, if successful response we're logged in and can view Story, if error, we're not logged in
Since there is no way for me to access the darth_key via JS (it's a 'httpOnly' cookie), this seemed like a way to go about it :/

Test Plan: Open a story in Viewmode, both logged in or not. If need be delete cookie _darth_key via browser devtools.

Reviewers: girish.vadlamudi, stephan.gensch

Reviewed By: girish.vadlamudi

Differential Revision: https://hive.vrgmnts.net/D1813